Make inReadOnlyTransaction part of SQLMetadataConnector
