+++ server/src/main/java/io/druid/metadata/SQLMetadataSegmentManager.java
@@ -160,44 +160,12 @@ public class SQLMetadataSegmentManager implements MetadataSegmentManager
     }
   }
 
-  protected static final <T> T inReadOnlyTransaction(
-      final SQLMetadataConnector connector,
-      final TransactionCallback<T> callback
-  )
-  {
-    return connector.getDBI().withHandle(
-        new HandleCallback<T>()
-        {
-          @Override
-          public T withHandle(Handle handle) throws Exception
-          {
-            final Connection connection = handle.getConnection();
-            final boolean readOnly = connection.isReadOnly();
-            connection.setReadOnly(true);
-            try {
-              return handle.inTransaction(callback);
-            }
-            finally {
-              try {
-                connection.setReadOnly(readOnly);
-              }
-              catch (SQLException e) {
-                // at least try to log it so we don't swallow exceptions
-                log.error(e, "Unable to reset connection read-only state");
-              }
-            }
-          }
-        }
-    );
-  }
-
   @Override
   public boolean enableDatasource(final String ds)
   {
     try {
       final IDBI dbi = connector.getDBI();
-      VersionedIntervalTimeline<String, DataSegment> segmentTimeline = inReadOnlyTransaction(
-          connector,
+      VersionedIntervalTimeline<String, DataSegment> segmentTimeline = connector.inReadOnlyTransaction(
           new TransactionCallback<VersionedIntervalTimeline<String, DataSegment>>()
           {
             @Override
@@ -479,8 +447,7 @@ public class SQLMetadataSegmentManager implements MetadataSegmentManager
       //
       // setting connection to read-only will allow some database such as MySQL
       // to automatically use read-only transaction mode, further optimizing the query
-      final List<DataSegment> segments = inReadOnlyTransaction(
-          connector,
+      final List<DataSegment> segments = connector.inReadOnlyTransaction(
           new TransactionCallback<List<DataSegment>>()
           {
             @Override
@@ -573,8 +540,7 @@ public class SQLMetadataSegmentManager implements MetadataSegmentManager
       final int limit
   )
   {
-    return inReadOnlyTransaction(
-        connector,
+    return connector.inReadOnlyTransaction(
         new TransactionCallback<List<Interval>>()
         {
           @Override
